+++ tests/e2e/append_object_test.go
@@ -30,7 +30,6 @@ var _ = ginkgo.Describe("Append objects", func() {
 	runAppendObjectSpecs()
 	runAppendCoalesceSpecs()
 	runAppendMidSizeBodySpecs()
-	runAppendForwardBufferSaturationSpecs()
 	runAppendSizeCapSpecs()
 })